Split main into main and options, make argv globally available
